implement sizes and uniform size for scatter
